pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mefai-signal-engine"
version = "1.0.0"
description = "Production-grade AI trading signal engine with XGBoost ensemble, multi-layer signal composition, and real-time data pipeline"
requires-python = ">=3.10"
dependencies = [
    "xgboost>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "apscheduler>=3.10.0",
    "aiohttp>=3.9.0",
    "feedparser>=6.0.0",
    "joblib>=1.3.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
]

[project.scripts]
signal-engine = "src.api.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
